feat: Migrate away from cloudflare
