+++ .vscode/settings.json
@@ -53,5 +53,6 @@
   },
   "typescript.tsdk": "node_modules/typescript/lib",
   "vitest.commandLine": "pnpm vitest",
-  "vitest.enable": true
+  "vitest.enable": true,
+  "cSpell.words": ["fastify", "Hono"]
 }
