--- packages/proxy-server/.gitignore
@@ -1,1 +1,0 @@
-.wrangler
