+++ packages/proxy-server/package.json
@@ -1,30 +1,34 @@
 {
   "name": "@scalar/proxy-server",
-  "description": "A simple proxy server running on hono which can be deployed anywhere!",
+  "description": "A simple proxy server which can be deployed anywhere!",
   "license": "MIT",
   "author": "Scalar (https://github.com/scalar)",
   "homepage": "https://github.com/scalar/scalar",
   "bugs": "https://github.com/scalar/scalar/issues/new/choose",
   "keywords": [
-    "api docs references typescript proxy cloudflare"
+    "api docs references typescript proxy"
   ],
   "version": "0.0.1",
   "engines": {
     "node": ">=18"
   },
   "scripts": {
-    "build": "tsc --project tsconfig.build.json",
-    "deploy": "wrangler deploy --minify src/proxy-worker.ts",
-    "dev:bun": "bun run --hot src/dev-server.ts",
-    "dev:cloudflare": "wrangler dev src/dev-server.ts --port 5051",
+    "build": "vite build && pnpm types:build && tsc-alias -p tsconfig.build.json",
+    "dev": "vite",
     "lint:check": "eslint .",
     "lint:fix": "eslint .  --fix",
+    "preview": "vite preview",
+    "test": "vitest",
+    "types:build": "tsc -p tsconfig.build.json",
     "types:check": "tsc --noEmit --skipLibCheck --composite false"
   },
   "type": "module",
   "main": "dist/index.js",
   "exports": {
-    "import": "./dist/index.js"
+    ".": {
+      "import": "./dist/index.js",
+      "types": "./dist/index.d.ts"
+    }
   },
   "files": [
     "dist",
@@ -37,7 +41,8 @@
     "directory": "packages/proxy-server"
   },
   "devDependencies": {
-    "@cloudflare/workers-types": "^4.20240208.0",
-    "wrangler": "^3.32.0"
+    "@scalar/build-tooling": "workspace:*",
+    "tsc-alias": "^1.8.8",
+    "vite": "^5.1.1"
   }
 }
