--- packages/proxy-server/src/dev-server.ts
@@ -1,6 +1,0 @@
-import proxyWorker from './proxy-worker'
-
-export default {
-  port: 5051,
-  fetch: proxyWorker.fetch,
-}
