+++ packages/proxy-server/src/index.ts
@@ -1,5 +1,1 @@
-import proxyServer from './proxy-worker'
-
-export const proxyFetch = proxyServer.fetch
-
-export default proxyServer
+export * from './proxy-fetch'
