+++ packages/proxy-server/tsconfig.json
@@ -1,8 +1,3 @@
 {
-  "extends": "../../tsconfig.json",
-  "compilerOptions": {
-    "lib": ["ESNext"],
-    "types": ["@cloudflare/workers-types"]
-  },
-  "outDir": "dist"
+  "extends": "../../tsconfig.json"
 }
